requirements.txt
flask
pygithub
python-dotenv
requests
openai
